Move Menu into containers
